dashboard files refactoring
