+++ src/app/components/dashboard/DashboardPanel.tsx
